Replace repo stuff with working Python3 version of the bot
This is a basic attendance points bot from long ago that still works.
It only requires a name and an oauth token. The host name is the name of whoever needs super permissions.
The memory file is stored in plaintext csv because I didn't trust sqlite at the time of writing.
To run, fill out the necessary info in config.ini and run chatbot.py.

Requires: irc, configparser, Python 3